# docs/api/openapi_effective_v0.2.3.yaml
...
openapi: 3.1.0
info:
  title: TURAB API
  version: 0.2.3
  description: 'TURAB foundational API contract v0.2 — remediated after Technical Architecture Review.


    Core invariants:

    - REQUEST and PROPERTY are equal first-class entities.

    - PROPERTY != POST/SOURCE and PROPERTY != PROPERTY_OFFER.

    - Match evaluation is Request × canonical Property plus an explicit commercial context.

    - MATCH_CANDIDATE != OPPORTUNITY.

    - Matching inputs are snapshotted immutably and hashed.

    - Hard FAIL cannot be overridden by ranking; required UNKNOWN creates information work.

    - Opportunity requires latest human APPROVED review plus freshness, permission and information gates.

    - Consent is resource-scoped and revocable.

    - Customer/public DTOs are separate from internal DTOs.

    - AI never upgrades verification, changes user criteria silently, or creates Opportunity autonomously.


    Technical Patch v0.2.3 is a metadata/version-stamp correction only; it preserves the v0.2.2 concurrency
    and PARTY-version semantics.

    '
servers:
- url: /api/v1
  description: Relative API base
...
paths:
...
  /requests:
    post:
      summary: Create request
      tags:
      - Requests
      security:
      - BearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '422': *id001
      x-roles:
      - ADMIN
      - OPERATOR
      - CUSTOMER
      operationId: postRequests
      x-authorization: Customer may act only on resources owned/managed by authenticated party; service
        MUST enforce object-level authorization in addition to role.
  /requests/{request_id}:
    get:
      summary: Get request
      tags:
      - Requests
      security:
      - BearerAuth: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema: *id002
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '422': *id001
      x-roles:
      - ADMIN
      - OPERATOR
      - REVIEWER
      operationId: getRequestsRequestId
      x-authorization: Internal resource access; role check plus business-purpose authorization. Customer
        access uses /me/* endpoints.
    patch:
      summary: Update request fields without silently relaxing criteria
      tags:
      - Requests
      security:
      - BearerAuth: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema: *id002
      - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '422': *id001
      x-roles:
      - ADMIN
      - OPERATOR
      - CUSTOMER
      operationId: patchRequestsRequestId
      description: Typed projection update only. Status, consent, claim ownership, availability and other
        stateful fields require dedicated domain commands. Critical fact updates MUST atomically write
        provenance/resolution plus the operational projection.
      x-authorization: Customer may act only on resources owned/managed by authenticated party; service
        MUST enforce object-level authorization in addition to role.
  /requests/{request_id}/criteria:
    post:
      summary: Add request criterion
      tags:
      - Requests
      security:
      - BearerAuth: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema: *id002
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCriterion'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestCriterion'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '422': *id001
      x-roles:
      - ADMIN
      - OPERATOR
      - CUSTOMER
      operationId: postRequestsRequestIdCriteria
      x-authorization: Customer may act only on resources owned/managed by authenticated party; service
        MUST enforce object-level authorization in addition to role.
  /requests/{request_id}/state:
    post:
      summary: Transition request lifecycle
      tags:
      - Requests
      security:
      - BearerAuth: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema: *id002
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestStateCommand'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '422': *id001
      x-roles:
      - ADMIN
      - OPERATOR
      - CUSTOMER
      operationId: postRequestsRequestIdState
      x-authorization: Customer may act only on resources owned/managed by authenticated party; service
        MUST enforce object-level authorization in addition to role.
      x-turab-workflow:
        id: CORRECTION-002
        decision: Slice 2 closing round, 2026-09-19
        approved: '2026-09-19'
        summary: 'ACTIVE -> PAUSED and ACTIVE -> CLOSED are adopted as direct transitions: a request is
          not required to pass through NEEDS_CONFIRMATION in order to be paused or closed. Reactivation
          from PAUSED or CLOSED is the state command itself with target_status=ACTIVE; no additional field
          is used or accepted. Closing requires a reason_code from the REQUEST_CLOSURE category added
          by migration 0002_request_closure_reasons, and REQUEST_CLOSED_OTHER requires a note.'
        reference: docs/gate/REQUEST_STATE_TRANSITIONS.md
  /requests/{request_id}/reconfirm:
    post:
      summary: Reconfirm request freshness
      tags:
      - Requests
      security:
      - BearerAuth: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema: *id002
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateReconfirm'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '422': *id001
      x-roles:
      - ADMIN
      - OPERATOR
      - CUSTOMER
      operationId: postRequestsRequestIdReconfirm
      x-authorization: Customer may act only on resources owned/managed by authenticated party; service
        MUST enforce object-level authorization in addition to role.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
      description: Required for mutating commands. Same key + same request hash must replay the original
        result; same key + different hash returns 409.
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema:
        type: integer
        minimum: 1
      description: Required integer optimistic-concurrency version for mutable resource projection.
...
  schemas:
...
    RequestCriterion:
      type: object
      required:
      - criterion_code
      - importance
      - operator
      - value
      properties:
        request_criterion_id: *id002
        criterion_code:
          type: string
        importance:
          type: string
          enum:
          - REQUIRED
          - PREFERRED
          - FLEXIBLE
        operator:
          type: string
          enum:
          - EQ
          - NEQ
          - IN
          - NOT_IN
          - GTE
          - LTE
          - BETWEEN
          - EXISTS
          - BOOL
          - TEXT_SEMANTIC
        value: {}
        unit:
          anyOf:
          - type: string
          - type: 'null'
        blocking_if_unknown:
          type: boolean
          default: false
        sort_order:
          type: integer
          default: 100
    RequestCreate:
      type: object
      required:
      - claim_status
      - intent
      - management_mode
      - party_id
      - transaction_intent
      properties:
        party_id: *id002
        intent: &id018
          type: string
          enum:
          - EXPLORING
          - ACTIVE_SEARCH
          - READY_TO_ACT
        payment: &id019
          type: string
          enum:
          - CASH
          - BANK_FINANCING
          - MIXED
          - UNDECIDED
        desired_property_type: &id020
          anyOf:
          - type: string
            enum:
            - HOUSE_VILLA
            - APARTMENT
            - LAND
            - SHOP_COMMERCIAL
            - AGRICULTURAL_PROPERTY
            - BUILDING
            - OTHER
          - type: 'null'
        property_type_importance:
          type: string
          enum:
          - REQUIRED
          - PREFERRED
          - FLEXIBLE
        primary_location_id: &id021
          anyOf:
          - *id002
          - type: 'null'
        location_importance:
          type: string
          enum:
          - REQUIRED
          - PREFERRED
          - FLEXIBLE
        local_location_detail:
          type: string
        budget_target_dzd: &id022
          type: integer
          minimum: 0
        budget_max_dzd: &id023
          type: integer
          minimum: 0
        budget_importance:
          type: string
          enum:
          - REQUIRED
          - PREFERRED
          - FLEXIBLE
        budget_flexibility: &id024
          type: string
          enum:
          - STRICT
          - LOW
          - MODERATE
          - HIGH
          - UNSPECIFIED
        management_mode:
          type: string
          enum:
          - SELF_MANAGED
          - ASSISTED
          - SHARED_MANAGEMENT
        criteria: &id026
          type: array
          items:
            $ref: '#/components/schemas/RequestCriterion'
        transaction_intent: &id017
          type: string
          enum:
          - BUY
          - RENT
        claim_status:
          type: string
          enum:
          - UNCLAIMED
          - CLAIMED
      additionalProperties: false
    Request:
      allOf:
      - $ref: '#/components/schemas/RequestCreate'
      - type: object
        required:
        - request_id
        - status
        - version
        - created_at
        properties:
          request_id: *id002
          status: &id016
            type: string
            enum:
            - RAW
            - CONTACTED
            - QUALIFIED
            - ACTIVE
            - NEEDS_CONFIRMATION
            - PAUSED
            - CLOSED
          last_confirmed_at: &id025
            anyOf:
            - *id003
            - type: 'null'
          claim_status:
            type: string
            enum:
            - UNCLAIMED
            - CLAIMED
          version:
            type: integer
            minimum: 1
          created_at: *id003
          updated_at: *id003
    RequestStateCommand:
      type: object
      required:
      - target_status
      properties:
        target_status:
          type: string
          enum:
          - CONTACTED
          - QUALIFIED
          - ACTIVE
          - NEEDS_CONFIRMATION
          - PAUSED
          - CLOSED
        reason_code:
          type: string
        note:
          type: string
      additionalProperties: false
...
    StateReconfirm:
      type: object
      properties:
        confirmed_at: *id003
        notes:
          type: string
      additionalProperties: false
...
    RequestPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        intent:
          type: string
          enum:
          - EXPLORING
          - ACTIVE_SEARCH
          - READY_TO_ACT
        payment:
          type: string
          enum:
          - CASH
          - BANK_FINANCING
          - MIXED
          - UNDECIDED
        desired_property_type:
          anyOf:
          - type: string
            enum:
            - HOUSE_VILLA
            - APARTMENT
            - LAND
            - SHOP_COMMERCIAL
            - AGRICULTURAL_PROPERTY
            - BUILDING
            - OTHER
          - type: 'null'
        primary_location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
        local_location_detail:
          type:
          - string
          - 'null'
        budget_target_dzd:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        budget_max_dzd:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        budget_flexibility:
          type: string
          enum:
          - STRICT
          - LOW
          - MODERATE
          - HIGH
          - UNSPECIFIED
...